+++ packages/bundle/src/adaptiveCards/Attachment/ThumbnailCardContent.tsx
@@ -2,7 +2,7 @@
 
 import { hooks } from 'botframework-webchat-component';
 import React, { memo, useMemo } from 'react';
-import { boolean, custom, object, optional, parse, pipe, readonly, safeParse, string, type InferInput } from 'valibot';
+import { boolean, object, optional, parse, pipe, readonly, string, type InferInput } from 'valibot';
 
 import useStyleOptions from '../../hooks/useStyleOptions';
 import useAdaptiveCardsPackage from '../hooks/useAdaptiveCardsPackage';
@@ -15,9 +15,7 @@ const { useDirection } = hooks;
 const thumbnailCardContentPropsSchema = pipe(
   object({
     actionPerformedClassName: optional(string()),
-    content: custom<InferInput<typeof directLineBasicCardSchema>>(
-      value => safeParse(directLineBasicCardSchema, value).success
-    ),
+    content: directLineBasicCardSchema,
     disabled: optional(boolean())
   }),
   readonly()
